requirements.txt
feedparser>=6.0.11
pyairtable>=2.3.3
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.3
schedule>=1.2.2
apify-client>=1.7.0
anthropic>=0.50.0
flask>=3.0.0
